Eliminate redundancy between init and main
